requirements.txt
streamlit==1.29.0
langchain==0.0.352
openai==1.6.1
python-dotenv==1.0.0
python-docx==1.0.1
pypdf2==3.0.1
pydantic==1.10.13
